# PERF: we don't need to use a huge image to test thumbnails