--- a/spec/models/topic_thumbnail_spec.rb
+++ b/spec/models/topic_thumbnail_spec.rb
@@ -2,15 +2,18 @@
 require 'rails_helper'
 
 describe "TopicThumbnail" do
-  let(:upload1) { Fabricate(:image_upload, width: 5000, height: 5000) }
+  let(:upload1) { Fabricate(:image_upload, width: 50, height: 50) }
   let(:topic) { Fabricate(:topic, image_upload: upload1) }
-  let(:upload2) { Fabricate(:image_upload, width: 5000, height: 5000) }
+  let(:upload2) { Fabricate(:image_upload, width: 50, height: 50) }
   let(:topic2) { Fabricate(:topic, image_upload: upload2) }
   let(:upload3) { Fabricate(:upload_no_dimensions) }
   let(:topic3) { Fabricate(:topic, image_upload: upload3) }
 
   before do
     SiteSetting.create_thumbnails = true
+
+    ImageSizer.stubs(:resize).returns([49, 49])
+
     topic.generate_thumbnails!(extra_sizes: nil)
 
     TopicThumbnail.ensure_consistency!
@@ -69,5 +72,4 @@ describe "TopicThumbnail" do
 
     expect(topic.topic_thumbnails.length).to eq(0)
   end
-
 end