@SamSaffron 2020-10-27
This is certainly a less fragile stub, I guess we could go in via plugin registry but I will let this one through for now :)